Added selector
