+++ package.json
@@ -32,6 +32,7 @@
     "redux": "^4.0.1",
     "redux-i18n": "^1.5.14",
     "redux-thunk": "^2.3.0",
+    "reselect": "^4.0.0",
     "underscore": "^1.9.1"
   },
   "devDependencies": {
